feat: adds crt and bubble shader
adds shader to give the effect of an older tv and
do the scanlines like a crt screen.

the bubble works great, but the crt has some issues
and this could be because of the way i am using s2d
in order to resize and scale the game.

i should test this and implement a different object
that does this so that s2d stays at the normal
resolution and position