fix: adds world object to manage placement
wow, i should have just done it, since that was
really quick and completely fixed the issue.

lesson learned: don't resize scenes ever, that will
cause issues with display, and shaders. instead
make an empty dumby object that will store everything
and then scale and resize as needed. (the world in
this case).