refactor: part one of restructure
restructuring some of this to make it easier to
adjust settings and build on it.

makes a new package `game` for all the game
specific logic and classes.

first part moves all the timer stuff to a new
clock object.

also moves the grid stuff to a new grid object
that manages positions and layer management.