chore: updates hxml ignores
changes the ignore string from build.hxml to *.hxml so there can be
a number of build files in the root of the project that are ignored
by git.

all the real build projects must be in /hxml. the reason for ignoring
them in the root is because the vscode tools can't use hxml files that
are in any other folder. i am using the root ones just as passthroughs
so they don't actually define anything else, just reference on of the
files in the hxml directory.