refactor: reimplements menu, splits settings
reimplements a menu system that should be easier to work with and
create more complex menus.

also splits the settings from one class `Settings` into a new package
with different classes based on what the setting is for. i could just
put these inside the classes they edit but i wanted a more centralized
space for them. this is an organizational experiment and i'll see if i
like this or prefer the older method. the idea is one space where all
the configuration options are.