fix: prevents player from moving into themselves
fixes a typo where you could make the player move backwards which would
cause deaths when trying to manuever at higher speeds. i originally
tried to solve this by looking at the moving direction and not the
direction that we will move at the next tick, but i didn't update one of
the references in the check causing it to be able to tricked if you used
the 'right' and 'down' buttons before trying to go backwards.