fix: blocking indications now works
fixes blocking indications, they didn't show after an update that i
didn't catch.

what was happening was that i changed the the display of the falls from
being colored as the food to being gray and outline with the food color.
i didn't catch that it was using the fill color to define the blocking
color (which because of the change was now gray).

so updated the blocking to check the outline color for the wall which
fixes the issue.

also adjusts some of the displaying so it lines up exactly over the
standard grid outlines (makes it look better).