fix: lcd shader now functions as expected
finally managed to understand and fix the lcd shader. the issue was that
i was assuming we were working on the entire screen resolution but
actually the shader was applied to `world` which is fixed and scaled per
the window resolution ... once i moved the shader from `world` to `s2d`
then the lcd shader was working as expected.

i had to rewrite lcd shader because if all the messing with it to get it
to work ... so not sure the exact changes in it, but it works now.

also this fixes the pixelation blockyness that is in the bubble shader,
so now it looks a lot smoother.