feat: adds palettes to the visual menu
makes a menu option to switch palettes in the game. was actually a huge ordeal because i had to go back to every graphical item and link it someway so that it would redraw and use the new colors when switching. so touches a lot of things in a lot of places.

codecks: $11z