fix: drawing issue with crtfilter & colors
when changing the background color to something other than black & when using the crt filter & bubble filter we can see pixels stop before hitting the edge of the canvas because of the way the shader is applied to the graphics.

the easiet fix (which this commit does) is to pin the background color (shader background color) to black when we we have the lcd effect enabled.