fix: fullscreen visual option respects actual state
the fullscreen option now actually checks if we are in fullscreen or not and sets the toggle to that state.

also adds more logic to the toggle so it checks if we actually should be toggling when we reset or now. so if the toggle is already in the "original state" then it won't do anything. 

codecks $124